Add ro.cm.device to build.prop
For market filtering reasons on certain devices (particularly
Galaxy S and its variants), we need to override ro.product.device
as market checks for that when doing device filtering (our build
tree and our proper ro.produce.device name are different). arcee
then informed me that it would break the mirroring system and CMStats
and suggested an alternative property. I would like to begin the
effort by adding "ro.cm.device", which defaults to the same value
as ro.product.device and is not affected by the override.

Change-Id: Ifec746e88a95a039462d3f5750cd07a5f055ab67